+++ dev/config/src/vite/index.ts
@@ -18,5 +18,7 @@ export { default as ViteFrontendConfig } from "./vite.frontend.config.js";
 export { default as ViteBackendConfig } from "./vite.backend.config.js";
 export { default as VitePluginFixAbsoluteImports } from "./vite-plugin-fix-absolute-imports.js";
 export { default as ViteCommonJSConfig } from "./vite.commonjs.config.js";
+export { default as ViteEsmConfig } from "./vite.esm.config.js";
 export { default as VitePluginRemoveUnusedTranslations } from "./vite-plugin-remove-unused-translations.js";
 export * from "./TsNoCheckPlugin.js";
+export { default as VitePluginCopy } from "./vite-plugin-copy.js";