+++ dev/scripts/src/setup/provider.ts
@@ -18,17 +18,9 @@ import { Tasks } from "@prosopo/provider";
 import type { IProviderAccount } from "@prosopo/types";
 import type { ProviderEnvironment } from "@prosopo/types-env";
 
-export async function setupProvider(
-	env: ProviderEnvironment,
-	provider: IProviderAccount,
-): Promise<void> {
-	if (!provider.pair) {
-		throw new ProsopoEnvError("DEVELOPER.MISSING_PROVIDER_PAIR", {
-			context: { provider },
-		});
-	}
+export async function setupProvider(env: ProviderEnvironment): Promise<void> {
 	const logger = env.logger;
 	const tasks = new Tasks(env);
-	logger.info("   - providerSetDataset");
+	logger.info(() => ({ msg: "   - providerSetDataset" }));
 	await tasks.datasetManager.providerSetDataset(datasetWithSolutionHashes);
 }