+++ packages/types-database/src/types/mongo.ts
@@ -20,6 +20,7 @@ export interface IDatabase {
 	dbname: string;
 	connection?: Connection;
 	logger: Logger;
+	connected: boolean;
 
 	getConnection(): Connection;