+++ packages/util-crypto/src/address/types.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2025 @polkadot/util-crypto authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+// FIXME we really want this to map with what is in the allowedSS58 array... i.e. the
+// values there. As of now, we just map to number.
+export type Prefix = number;