+++ packages/types-database/CHANGELOG.md
@@ -1,5 +1,230 @@
 # @prosopo/types-database
 
+## 3.0.16
+### Patch Changes
+
+- 1f3a02f: Release 3.1.8
+- Updated dependencies [1f3a02f]
+  - @prosopo/user-access-policy@3.5.1
+
+## 3.0.15
+### Patch Changes
+
+- Updated dependencies [e0628d9]
+  - @prosopo/user-access-policy@3.5.0
+
+## 3.0.14
+### Patch Changes
+
+- Updated dependencies [44ffda2]
+- Updated dependencies [a49b538]
+- Updated dependencies [e090e2f]
+  - @prosopo/config@3.1.4
+  - @prosopo/user-access-policy@3.4.1
+  - @prosopo/common@3.1.3
+  - @prosopo/types@3.0.7
+
+## 3.0.13
+### Patch Changes
+
+- 828066d: remove empty test npm scripts, add missing npm test scripts
+- 91bbe87: configure typecheck before bundle for vue packages
+- 91bbe87: make typecheck script always recompile
+- 346e092: NODE_ENV default to "development"
+- 5d36e05: remove tsc --force
+- Updated dependencies [828066d]
+- Updated dependencies [df4e030]
+- Updated dependencies [91bbe87]
+- Updated dependencies [3ef4fd2]
+- Updated dependencies [91bbe87]
+- Updated dependencies [346e092]
+- Updated dependencies [5d36e05]
+  - @prosopo/common@3.1.2
+  - @prosopo/types@3.0.6
+  - @prosopo/config@3.1.3
+  - @prosopo/user-access-policy@3.4.0
+
+## 3.0.12
+### Patch Changes
+
+- eb71691: configure typecheck before bundle for vue packages
+- eb71691: make typecheck script always recompile
+- Updated dependencies [eb71691]
+- Updated dependencies [eb71691]
+  - @prosopo/user-access-policy@3.3.2
+  - @prosopo/common@3.1.1
+  - @prosopo/types@3.0.5
+  - @prosopo/config@3.1.2
+
+## 3.0.11
+### Patch Changes
+
+- 625fef8: ua parsing
+
+## 3.0.10
+### Patch Changes
+
+- 3573f0b: fix npm scripts bundle command
+- 3573f0b: build using vite, typecheck using tsc
+- efd8102: Add tests for unwrap error helper
+- 3573f0b: standardise all vite based npm scripts for bundling
+- Updated dependencies [93d5e50]
+- Updated dependencies [3573f0b]
+- Updated dependencies [3573f0b]
+- Updated dependencies [efd8102]
+- Updated dependencies [93d5e50]
+- Updated dependencies [63519d7]
+- Updated dependencies [f29fc7e]
+- Updated dependencies [3573f0b]
+- Updated dependencies [2d0dd8a]
+  - @prosopo/types@3.0.4
+  - @prosopo/user-access-policy@3.3.1
+  - @prosopo/common@3.1.0
+  - @prosopo/config@3.1.1
+
+## 3.0.9
+### Patch Changes
+
+- Updated dependencies [b7c3258]
+  - @prosopo/user-access-policy@3.3.0
+
+## 3.0.8
+### Patch Changes
+
+- Updated dependencies [cdf7c29]
+  - @prosopo/user-access-policy@3.2.1
+
+## 3.0.7
+### Patch Changes
+
+- Updated dependencies [a7164ce]
+  - @prosopo/user-access-policy@3.2.0
+
+## 3.0.6
+### Patch Changes
+
+- b0d7207: Types for proper rotation
+- Updated dependencies [b0d7207]
+  - @prosopo/types@3.0.3
+  - @prosopo/user-access-policy@3.1.5
+
+## 3.0.5
+### Patch Changes
+
+  - @prosopo/user-access-policy@3.1.4
+
+## 3.0.4
+### Patch Changes
+
+  - @prosopo/user-access-policy@3.1.3
+
+## 3.0.3
+### Patch Changes
+
+- f682f0c: Moving type and fixing i18n config
+- Updated dependencies [f682f0c]
+  - @prosopo/types@3.0.2
+  - @prosopo/common@3.0.2
+  - @prosopo/user-access-policy@3.1.2
+
+## 3.0.2
+### Patch Changes
+
+  - @prosopo/common@3.0.1
+  - @prosopo/types@3.0.1
+  - @prosopo/user-access-policy@3.1.1
+
+## 3.0.1
+### Patch Changes
+
+- 913f2a6: Make custom expiration times work in redis. Make redis internal only and persist data
+- Updated dependencies [913f2a6]
+  - @prosopo/user-access-policy@3.1.0
+
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/user-access-policy@3.0.0
+  - @prosopo/common@3.0.0
+  - @prosopo/types@3.0.0
+
+## 2.7.6
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+  - @prosopo/user-access-policy@2.6.8
+
+## 2.7.5
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/user-access-policy@2.6.7
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+
+## 2.7.4
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/types@2.9.0
+  - @prosopo/common@2.7.1
+  - @prosopo/user-access-policy@2.6.6
+
+## 2.7.3
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/user-access-policy@2.6.5
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/types@2.7.1
+  - @prosopo/user-access-policy@2.6.4
+
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies [6e1aef6]
+  - @prosopo/types@2.7.0
+  - @prosopo/user-access-policy@2.6.3
+
+## 2.7.0
+
+### Minor Changes
+
+- cf59998: Update DB schema
+
+## 2.6.2
+
+### Patch Changes
+
+- Updated dependencies [6ff193a]
+  - @prosopo/types@2.6.2
+  - @prosopo/user-access-policy@2.6.2
+
+## 2.6.1
+
+### Patch Changes
+
+- 52feffc: Adjustable difficulty img captcha
+- Updated dependencies [52feffc]
+  - @prosopo/types@2.6.1
+  - @prosopo/user-access-policy@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes